+++ android/core/engine/src/main/java/com/xianxia/sect/core/engine/domain/economy/EconomyFacade.kt
@@ -3,7 +3,6 @@ package com.xianxia.sect.core.engine.domain.economy
 import com.xianxia.sect.core.engine.domain.inventory.InventoryFacade
 import com.xianxia.sect.core.engine.domain.save.SaveFacade
 import com.xianxia.sect.core.engine.service.AutoBuyService
-import com.xianxia.sect.core.engine.service.DailySignInService
 import com.xianxia.sect.core.engine.service.MailService
 import com.xianxia.sect.core.engine.service.RedeemCodeService
 import com.xianxia.sect.core.wallet.SpiritStoneWallet
@@ -18,6 +17,5 @@ interface EconomyFacade {
     val autoBuyService: AutoBuyService
     val redeemCodeService: RedeemCodeService
     val mailService: MailService
-    val dailySignInService: DailySignInService
     val saveFacade: SaveFacade
 }